feat: bootstrap app
