+++ lib/app/view/app.dart
@@ -1,0 +1,23 @@
+import 'package:ax_dapp/pages/landing_page/landing_page.dart';
+import 'package:flutter/material.dart';
+
+class App extends StatelessWidget {
+  const App({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'AthleteX',
+      debugShowCheckedModeBanner: false,
+      initialRoute: '/',
+      theme: ThemeData(
+        canvasColor: Colors.transparent,
+        brightness: Brightness.dark,
+        primaryColor: Colors.yellow[700],
+        colorScheme: ColorScheme.fromSwatch(brightness: Brightness.dark)
+            .copyWith(secondary: Colors.black),
+      ),
+      home: const LandingPage(),
+    );
+  }
+}
