+++ lib/main.dart
@@ -1,6 +1,8 @@
+import 'dart:async';
 import 'dart:developer';
 
-import 'package:ax_dapp/pages/landing_page/landing_page.dart';
+import 'package:ax_dapp/app/app.dart';
+import 'package:ax_dapp/bootstrap.dart';
 import 'package:ax_dapp/repositories/coin_gecko_repo.dart';
 import 'package:ax_dapp/repositories/mlb_repo.dart';
 import 'package:ax_dapp/repositories/nfl_repo.dart';
@@ -16,7 +18,6 @@ import 'package:ax_dapp/service/graphql/graphql_client_helper.dart';
 import 'package:ax_dapp/service/graphql/graphql_configuration.dart';
 import 'package:coingecko_api/coingecko_api.dart';
 import 'package:dio/dio.dart';
-import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:graphql_flutter/graphql_flutter.dart';
 
@@ -31,61 +32,40 @@ void main() async {
   final _getPairInfoUseCase = GetPairInfoUseCase(_subGraphRepo);
   final _getSwapInfoUseCase = GetSwapInfoUseCase(_getPairInfoUseCase);
   log('GraphQL Client initialized}');
-  runApp(
-    GraphQLProvider(
-      client: _gQLClient,
-      child: MultiRepositoryProvider(
-        providers: [
-          RepositoryProvider(create: (context) => _subGraphRepo),
-          RepositoryProvider(
-            create: (context) => MLBRepo(_mlbApi),
-          ),
-          RepositoryProvider(
-            create: (context) => NFLRepo(),
-          ),
-          RepositoryProvider(
-            create: (context) => CoinGeckoRepo(_coinGeckoApi),
-          ),
-          RepositoryProvider(create: (context) => _getPairInfoUseCase),
-          RepositoryProvider(create: (context) => _getSwapInfoUseCase),
-          RepositoryProvider(
-            create: (context) => GetBuyInfoUseCase(_getSwapInfoUseCase),
-          ),
-          RepositoryProvider(
-            create: (context) => GetSellInfoUseCase(_getSwapInfoUseCase),
-          ),
-          RepositoryProvider(
-            create: (context) => GetPoolInfoUseCase(_getPairInfoUseCase),
-          ),
-          RepositoryProvider(
-            create: (context) => GetAllLiquidityInfoUseCase(_subGraphRepo),
-          ),
-        ],
-        child: const MyApp(),
+  unawaited(
+    bootstrap(
+      () => GraphQLProvider(
+        client: _gQLClient,
+        child: MultiRepositoryProvider(
+          providers: [
+            RepositoryProvider(create: (context) => _subGraphRepo),
+            RepositoryProvider(
+              create: (context) => MLBRepo(_mlbApi),
+            ),
+            RepositoryProvider(
+              create: (context) => NFLRepo(),
+            ),
+            RepositoryProvider(
+              create: (context) => CoinGeckoRepo(_coinGeckoApi),
+            ),
+            RepositoryProvider(create: (context) => _getPairInfoUseCase),
+            RepositoryProvider(create: (context) => _getSwapInfoUseCase),
+            RepositoryProvider(
+              create: (context) => GetBuyInfoUseCase(_getSwapInfoUseCase),
+            ),
+            RepositoryProvider(
+              create: (context) => GetSellInfoUseCase(_getSwapInfoUseCase),
+            ),
+            RepositoryProvider(
+              create: (context) => GetPoolInfoUseCase(_getPairInfoUseCase),
+            ),
+            RepositoryProvider(
+              create: (context) => GetAllLiquidityInfoUseCase(_subGraphRepo),
+            ),
+          ],
+          child: const App(),
+        ),
       ),
     ),
   );
-}
-
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
-
-  // This widget is the root of your application.
-  @override
-  Widget build(BuildContext context) {
-    // Returns anything!
-    return MaterialApp(
-      title: 'AthleteX',
-      debugShowCheckedModeBanner: false,
-      initialRoute: '/',
-      theme: ThemeData(
-        canvasColor: Colors.transparent,
-        brightness: Brightness.dark,
-        primaryColor: Colors.yellow[700],
-        colorScheme: ColorScheme.fromSwatch(brightness: Brightness.dark)
-            .copyWith(secondary: Colors.black),
-      ),
-      home: const LandingPage(),
-    );
-  }
 }
